#164 Restructure the test folder
